Remove HTMLFrameElement.location
Intent to Deprecate and Remove:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/aLpDg8pviU0/dUbUt7jsvT4J

javascriptURL-execution-context-frame-location-htmldom.html was
identical to javascriptURL-execution-context-frame-src-htmldom.html
apart from the attribute set, so remove it.

javascriptURL-execution-context-frame-location-htmldom-uncaught.html
tests code that was removed in V8HTMLFrameElementCustom.cpp and does not
have a counterpart for the src attribute, so remove it too.

BUG=358505

Review URL: https://codereview.chromium.org/220563002

git-svn-id: svn://svn.chromium.org/blink/trunk@170563 bbb929c8-8fbe-4397-9dbb-9b2b20218538