Conversion of more places in Blink to use WebLocalFrame
I'm breaking up my local mega patch to land in several chunks. This is
the Blink portion that's needed to build blink_tests successfully with
WebLocalFrame subclassing WebFrame.

Overall, the changes are pretty straightforward. One method changed back
to taking a WebFrame*. It's not yet clear how we're going to coordinate
frame detach for remote frames. Once there's a clear story, I'll update
it appropriately.

The typedef remains for now, as I work on landing the Chrome bits that
depend on this.

BUG=346764
R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/222203007

git-svn-id: svn://svn.chromium.org/blink/trunk@170717 bbb929c8-8fbe-4397-9dbb-9b2b20218538