Don't repaint when setting border or outline 0 with style none
We were causing repaints when setting the width of border or outline to
0 on an element that has outline or border style of none because the
internal default width is 3px, even though we compute the width to 0
at layout and paint time.

This patch adds a new method visuallyEqual and uses that when doing the
visualInvalidationDiff to avoid causing repaints when the visual output
would be identical.

It also apparently also reduces the amount of repainting in
fast/repaint/table-collapsed-border.html the table tests, but the visual
output is identical so it appears correct.

BUG=357629

Review URL: https://codereview.chromium.org/222473005

git-svn-id: svn://svn.chromium.org/blink/trunk@170743 bbb929c8-8fbe-4397-9dbb-9b2b20218538