Fix RenderLayer::hasNonCompositedChild
Another case of updating a tree order property while doing a paint order
walk. Since the neg z-order descendants are processed before the parent,
the code for clearing the bit on the parent clobbered the data from the
children.

This CL updates the property in the same way that
hasBlockSelectionGapBounds is (correctly) updated.

BUG=358877

Review URL: https://codereview.chromium.org/225173003

git-svn-id: svn://svn.chromium.org/blink/trunk@170888 bbb929c8-8fbe-4397-9dbb-9b2b20218538