Performance Test for Many Containers with Floats and a Normal Div
This is a test for https://codereview.chromium.org/204353009/ to
show progress against when it lands. It consists
of a series of 1000 siblings, each containing 10 floats
followed by a normal flow div. During layout about 3% of
the time is spent iterating back through the previous
sibling floats when laying out the normal flow child.

Review URL: https://codereview.chromium.org/227573006

git-svn-id: svn://svn.chromium.org/blink/trunk@170993 bbb929c8-8fbe-4397-9dbb-9b2b20218538