Measure usage for Selection.setPosition
Selection.setPosition was there from the very beginning:
http://src.chromium.org/viewvc/blink/trunk/WebCore/bindings/js/kjs_window.cpp?r1=6266&r2=6314&pathrev=24000

Currently it is just an alias of Selection.collapse function.
The setPosition function is a non standard function and both IE11 and FireFox don't support the function.
We should add a deprecation warning. 
Usage statistics would allow us to determine if it can be deprecated.

BUG=361002

Review URL: https://codereview.chromium.org/227493002

git-svn-id: svn://svn.chromium.org/blink/trunk@171035 bbb929c8-8fbe-4397-9dbb-9b2b20218538