Web Animations: Move LengthUnitType enum out of AnimatableLength and into CSSPrimitiveValue
This patch moves the LengthUnitType enum into CSSPrimitiveValue along with
converters between it and the CSSPrimitiveValue::UnitTypes enum.
This enum is intended to also be used by other parts of the animation engine
to communicate with CSSCalcValues about the type of length units they hold.
There is no change in functionality as a result of this patch.

Review URL: https://codereview.chromium.org/224723015

git-svn-id: svn://svn.chromium.org/blink/trunk@171126 bbb929c8-8fbe-4397-9dbb-9b2b20218538