Early return in RenderLayerCompositor::updateForceCompositingMode.
We're hitting an assert in requiresCompositingForScrollableFrame,
because it depends on layout information. With
http://src.chromium.org/viewvc/blink?revision=171366&view=revision,
we now try to update this state when setting forceCompositingMode,
at which point, layout is not necessarily up to date.

We can just early return here because this method will be called
again after we've done the layout. This does potentially expose
us to bugs from chicken-egg issues that depend on the
m_forceCompositingMode bit for subframes.

Review URL: https://codereview.chromium.org/236193002

git-svn-id: svn://svn.chromium.org/blink/trunk@171389 bbb929c8-8fbe-4397-9dbb-9b2b20218538