Fix positioning of link highlights for squashed layers.
Previously, these had been positioned wrt the grouped mapping for the
squashed layer. This was wrong. The squashing layer (m_squashingLayer)
is often offset from the CLM. The result is that the link highlight
does not accout for the position of the m_squashingLayer and it's not
aligned with the stuff it's supposed to highlight.

R=abarth@chromium.org
BUG=351056

Review URL: https://codereview.chromium.org/237763003

git-svn-id: svn://svn.chromium.org/blink/trunk@171518 bbb929c8-8fbe-4397-9dbb-9b2b20218538